Marked classes working with in-memory database as Deprecated, added configuration for Hibernate
