+++ pom.xml
@@ -6,18 +6,25 @@
 
     <groupId>com.epam.wca.gym</groupId>
     <artifactId>GymSpringCore</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <version>1.1-SNAPSHOT</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <java.version>17</java.version>
-        <lombok.version>1.18.30</lombok.version>
+
         <spring.version>6.1.6</spring.version>
         <javax.annotation.version>1.3.2</javax.annotation.version>
-        <opencsv.version>5.9</opencsv.version>
+
+        <lombok.version>1.18.30</lombok.version>
+
+        <java.persistence.api.version>2.2</java.persistence.api.version>
+        <postgresql.version>42.7.4</postgresql.version>
+        <hibernate.core.version>6.6.0.Final</hibernate.core.version>
+
         <slf4j.version>2.0.13</slf4j.version>
         <logback.version>1.4.14</logback.version>
+
         <junit.jupiter.version>5.10.2</junit.jupiter.version>
         <junit.jupiter.params.version>5.10.2</junit.jupiter.params.version>
         <junit.jupiter.engine.version>5.10.2</junit.jupiter.engine.version>
@@ -29,6 +36,8 @@
     </properties>
 
     <dependencies>
+
+        <!-- Spring Dependencies -->
         <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-context</artifactId>
@@ -39,12 +48,33 @@
             <artifactId>javax.annotation-api</artifactId>
             <version>${javax.annotation.version}</version>
         </dependency>
+
+        <!-- Database Dependencies -->
+        <dependency>
+            <groupId>javax.persistence</groupId>
+            <artifactId>javax.persistence-api</artifactId>
+            <version>${java.persistence.api.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>postgresql</artifactId>
+            <version>${postgresql.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.hibernate</groupId>
+            <artifactId>hibernate-core</artifactId>
+            <version>${hibernate.core.version}</version>
+        </dependency>
+
+        <!-- Utility Dependency -->
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
             <version>${lombok.version}</version>
             <scope>provided</scope>
         </dependency>
+
+        <!-- Logging Dependencies -->
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
@@ -55,11 +85,8 @@
             <artifactId>logback-classic</artifactId>
             <version>${logback.version}</version>
         </dependency>
-        <dependency>
-            <groupId>com.opencsv</groupId>
-            <artifactId>opencsv</artifactId>
-            <version>${opencsv.version}</version>
-        </dependency>
+
+        <!-- Testing Dependencies -->
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-api</artifactId>
@@ -83,7 +110,6 @@
             <artifactId>junit-platform-commons</artifactId>
             <version>${junit.platrofm.commons.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-junit-jupiter</artifactId>
@@ -105,8 +131,8 @@
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>${maven.compiler.plugin.version}</version>
                 <configuration>
-                    <source>17</source>
-                    <target>17</target>
+                    <source>${java.version}</source>
+                    <target>${java.version}</target>
                 </configuration>
             </plugin>
             <plugin>
