+++ src/main/java/com/epam/wca/gym/facade/MainFacade.java
@@ -1,19 +1,9 @@
 package com.epam.wca.gym.facade;
 
-import com.epam.wca.gym.config.StorageConfig;
-import com.epam.wca.gym.repository.impl.TraineeDAOImpl;
-import com.epam.wca.gym.repository.impl.TrainerDAOImpl;
-import com.epam.wca.gym.repository.impl.TrainingDAOImpl;
-import com.epam.wca.gym.service.impl.ProfileServiceImpl;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-
-import java.util.Arrays;
-
 public class MainFacade {
 
     public static void main(String[] args) {
-
+/*
         ApplicationContext context = new AnnotationConfigApplicationContext(StorageConfig.class);
 
         TraineeDAOImpl traineeDAO = context.getBean(TraineeDAOImpl.class);
@@ -39,6 +29,8 @@ public class MainFacade {
 
         ProfileServiceImpl profileService = context.getBean(ProfileServiceImpl.class);
         System.out.println(profileService.getUsernameCounter());
+
+ */
 
     }
 }
