+++ src/main/java/com/epam/wca/gym/service/in_memory_impl/ProfileServiceImpl.java
@@ -1,22 +1,35 @@
-package com.epam.wca.gym.service.impl;
+package com.epam.wca.gym.service.in_memory_impl;
 
 import com.epam.wca.gym.entity.Trainee;
 import com.epam.wca.gym.entity.Trainer;
 import com.epam.wca.gym.service.ProfileService;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
 import java.security.SecureRandom;
 import java.util.Map;
 
+/**
+ * Implementation of {@link ProfileService} that provides user profile management functionalities,
+ * including username and password generation.
+ *
+ * <p>This service is responsible for creating unique usernames and generating random passwords for users.</p>
+ *
+ * @deprecated This class is deprecated and will be removed in future versions. It is recommended to use
+ * an updated implementation that supports additional security features and integrates with the new database
+ * architecture. Future updates will include enhanced password hashing and a more scalable username management
+ * system.
+ * @since 1.1
+ * @see ProfileService
+ */
+
+@Deprecated(since = "1.1", forRemoval = false)
+
 @Getter
 @Slf4j
-@Service
 public class ProfileServiceImpl implements ProfileService {
-    @Autowired
+//    @Autowired
     private Map<String, Integer> usernameCounter;
     private static final String CHARACTERS = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
     private static final int PASSWORD_LENGTH = 10;
