+++ src/main/java/com/epam/wca/gym/util/HibernateUtil.java
@@ -1,0 +1,28 @@
+package com.epam.wca.gym.util;
+
+import lombok.experimental.UtilityClass;
+import org.hibernate.SessionFactory;
+import org.hibernate.cfg.Configuration;
+
+@UtilityClass
+public class HibernateUtil {
+    private static final SessionFactory sessionFactory;
+
+    static {
+        try {
+            // Load the hibernate.cfg.xml configuration
+            sessionFactory = new Configuration().configure().buildSessionFactory();
+        } catch (Exception ex) {
+            throw new ExceptionInInitializerError(ex);
+        }
+    }
+
+    public static SessionFactory getSessionFactory() {
+        return sessionFactory;
+    }
+
+    public static void shutdown() {
+        getSessionFactory().close();
+    }
+}
+
