+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-data.file.path = classpath:data/DataFileDefault.csv
+#data.file.path = classpath:data/DataFileDefault.csv
