+++ src/test/java/com/epam/wca/gym/service/ProfileServiceImplTest.java
@@ -1,6 +1,6 @@
 package com.epam.wca.gym.service;
 
-import com.epam.wca.gym.service.impl.ProfileServiceImpl;
+import com.epam.wca.gym.service.in_memory_impl.ProfileServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
